+++ buildout/src/Euphorie/docs/changes.rst
@@ -4,6 +4,11 @@ Changelog
 2.2 - Unreleased
 ----------------
 
+- Do not accept pre-1900 dates in the action plan, since they break rendering of the report.
+  This prevents `TNO ticket 150
+  <http://code.simplon.biz/tracker/tno-euphorie/ticket/150>`_ from occuring.
+  [wichert]
+
 - Fixed the logo markup according to Cornelis' requirements to match the final OiRA logo
   [pilz]